localmarket_backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
python-decouple>=3.8

